[AdaptiveMutationOperator] Implementation, mutation can be profiled for each species. Right now mostly for lifespan, but will surely change.
