+++ biome/systems/evolution/genes/fauna_genes.py
@@ -17,7 +17,10 @@
 """
 from typing import List, Dict, Any
 
-from biome.systems.evolution.genes.genes import Genes
+from deap import creator
+
+from biome.entities.entity import Entity
+from biome.systems.evolution.genes.genes import Genes, extract_common_genes
 from shared.enums.enums import ComponentType
 
 
@@ -95,3 +98,59 @@ class FaunaGenes(Genes):
     @lifespan.setter
     def lifespan(self, value: float):
         self._lifespan = value
+
+
+def fauna_genes_to_individual(fauna_genes: FaunaGenes):
+    fauna_genes.validate_genes()
+
+    return creator.Individual([
+        (fauna_genes.growth_modifier - 0.1) / 1.9,  # De 0.1-2.0 a 0-1
+        (fauna_genes.growth_efficiency - 0.1) / 0.9,  # De 0.1-1.0 a 0-1
+        (fauna_genes.max_size - 0.1) / 4.9,  # De 0.1-5.0 a 0-1
+        (fauna_genes.max_vitality - 50.0) / 150.0,  # De 50-200 a 0-1
+        (fauna_genes.aging_rate - 0.1) / 1.9,  # De 0.1-2.0 a 0-1
+        (fauna_genes.health_modifier - 0.1) / 1.9,  # De 0.1-2.0 a 0-1
+        (fauna_genes.max_energy_reserves - 50.0) / 100.0,  # De 50-150 a 0-1
+        (fauna_genes.lifespan - 1.0) / 999.0,  # De 1-1000 a 0-1
+
+        (fauna_genes.cold_resistance - 0.01) / 0.99,  # 0-1
+        (fauna_genes.heat_resistance - 0.01) / 0.99,  # 0-1
+
+        (fauna_genes.optimal_temperature + 20.0) / 60.0,  # De -20-40 a [0,1]
+
+        # TODO: Especificos de fauna cuando  haga
+        # (fauna_genes.foraging_efficiency - 0.1) / 0.9,  # De 0.1-1.0 a 0-1
+        # fauna_genes.predator_avoidance,  # De 0.0-1.0 a 0-1
+    ])
+
+
+def deap_genes_to_fauna_genes(individual) -> FaunaGenes:
+    genes = FaunaGenes()
+
+    clamped_individual = [max(0.0, min(1.0, val)) for val in individual]
+
+    genes.growth_modifier = 0.1 + (clamped_individual[0] * 1.9)  # 0.1-2.0
+    genes.growth_efficiency = 0.1 + (clamped_individual[1] * 0.9)  # 0.1-1.0
+    genes.max_size = 0.1 + (clamped_individual[2] * 4.9)  # 0.1-5.0
+    genes.max_vitality = 50.0 + (clamped_individual[3] * 150.0)  # 50-200
+    genes.aging_rate = 0.1 + (clamped_individual[4] * 1.9)  # 0.1-2.0
+    genes.health_modifier = 0.1 + (clamped_individual[5] * 1.9)  # 0.1-2.0
+    genes.max_energy_reserves = 50.0 + (clamped_individual[6] * 100.0)  # 50-150
+    genes.lifespan = 1.0 + (clamped_individual[7] * 999.0)  # 1-1000 años
+
+    genes.cold_resistance = 0.01 + clamped_individual[8] * 0.99  # 0-1
+    genes.heat_resistance = 0.01 + clamped_individual[9] * 0.99  # 0-1
+
+    genes.optimal_temperature = -20.0 + (clamped_individual[10] * 60.0)  # -20 a 40
+
+    # TODO: Especificos de fauna cuando  haga
+    # genes.foraging_efficiency = 0.1 + (clamped_individual[12] * 0.9)  # 0.1-1.0
+    # genes.predator_avoidance = clamped_individual[13]  # 0.0-1.0
+
+    return genes
+
+def extract_genes_from_fauna(fauna_entity: Entity) -> FaunaGenes:
+    genes = FaunaGenes()
+    extract_common_genes(fauna_entity, genes)
+    # TODO: Específicos de fauna, cuando los tenga
+    return genes
